mv wb.catalog._schema to model_library.api_web catalog
